+++ cookbooks/chub-vault/recipes/default.rb
@@ -10,6 +10,16 @@
 include_recipe "apt"
 include_recipe "java"
 
+# Host file tweaks needed for remote JMX to work
+hostsfile_entry '127.0.1.1' do
+  action    :remove
+end
+hostsfile_entry node['ipaddress'] do
+  hostname  node['fqdn']
+  aliases   [node['hostname']]
+  action    :create
+end
+
 group "vault" do
     action :create
     system true